Update README.md
extensions of the core of a simple functional
crown, which we will call tinyos-3. Unlike the operating systems we use,
our system will not run directly on the system hardware, but as a process of a
of another operating system, Linux, will use simulated resources, i.e. it will
it is written in a (simple) virtual machine. That is, instead of tinyos-3 managing the material resources of one
computer, it will manage the resources (CPU time, memory, peripherals, etc.) given to it by
the foreign OS (host OS). Apart from this difference, tinyos-3 will perform the basic functions of one
real operating system.